Change OutputId::get() to return u64

+++ src/backend/mod.rs
@@ -44,8 +44,8 @@ impl OutputId {
         OutputId(OUTPUT_ID_COUNTER.next())
     }
 
-    pub fn get(self) -> u32 {
-        self.0
+    pub fn get(self) -> u64 {
+        u64::from(self.0)
     }
 }
